working on recovery system
in effect, making a policy system that requires some combination of
signatures on a recovery request (ideally, these signatures would come
from other well-establish identities) that allow replacing the recovery
key with a new one.

this required changing the key structure. where before the recovery key
was on top, now we have an alpha key, which controls both the policy and
recovery keys, and then our normal root signing key (under the recovery
key) and our subkeys.

i'm still not entirely 100% on the organization of it all (like, ming
end up putting the policy key above the recovery key since they are
siblings) but in all it's starting to take shape.