Keychain and DAG stuff
- rewriting a bunch of keychain crap. remove KeyID as a signature ID
  from subkeys, and adding a new KeyID object which is effectively a
  wrapper around a public key
- subkeys are now referenced by their name field when
  revoking/deleting/etc (they no longer have ids, remember?)
- implementing all of the outer transaction functions for the DAG and
  some of the inner (ie, `apply_transaction()`) implementations. making
  huge progress here
- fleshing out a lot of the recovery system (kind of a rewrite). the DAG
  has made it almost impossibly simple to work with. there's still a lot
  to do, but it comes down to actual verification of the
  policy/requests, as opposed to having some weird, shitty, half-baked
  recovery transaction log like i was trying to do before.

overall, the DAG has simplified a bunch of stuff and i've been going
through the project with a medium-tooth comb rewriting things that no
longer make sense. i think we're at the point where like 90% of it is in
a good spot, and the rest will come moving forwrd. i want to really sit
down and focus on the unit tests for the identity/dag system, then work
on the recovery policy junk. there's a lot of hidden complexity in the
recovery...it's much simpler now than it was, but there is much to do.