replacing sodox::sign::* with ed25519-dalek
not super happy about this, because the two are NOT compatible, either
in key generation from seed or even in valid signature results (sigs
from sodox are "invalid" in ed25519-dalek for some godforsaken reason).

basically, this means all existing identities are completely invalid and
need to be regenerated, which sucks because i have to rebuild the Zephy
ID in the stamp doc (lol).

also, had to update a shit ton of tests and fix some serialization junk.
also, implementing rand_chacha20 since i thought it would make the
derive-from-seed signing keys line up. nope. oh well, at least it's more
secure than the StdRnd or w/e the hell it's called.