Converting all binary serialization to ASN.1
this one was a fucking doozy. ripped out all msgpack and replaced with
ASN.1 DER (via the wonderful and capable rasn crate).

this wasn't just adding a bunch of dumb #[derive(...)] macros, but
actually changing the fundamental structure of a lot of enums. this was
because multi-tuple enum variants (ie `MyEnum::Keypair(public, secret)`)
cannot be serialized via the current derive macros for rasn (or, for ANY
ASN.1 implementation to be fair). instead of adding it, instead i
decided to just refactor those variants into named struct enum variants.
this is much more clear, if not a bit more verbose.

this obviously changed how signatures are created because ANS.1 =/=
MsgPack so a lot of the tests had to change.

overall, this is for the better. i ignored ASN.1 for years because i
thought it was some antiquated format, but it's purpose-built for
EXACTLY our use-case and much more expressive than any of the other
binary serialization formats (protobufs, msgpack, etc). i can't believe
more people don't use it. WTF.

anyway, tests passing (including some new ones to head some issues that
came up during the conversion off at the pass).

what a wild ride.