A checkin on stamp-protocol/tracker#31
Transaction IDs are onw a sha512 of the entry and allow multiple sigs.

Policies are used for all tranactions (unimplemented, but the structure
is there).

Names for keys are no longer unique...keys are handled by id, and
SecretKeys now have ids (their hmac).

all private containers are PrivateWithHmac now, meaning their content
can be verified AND their HMAC can be included in transactions.

adding names to claims (again, not unique). this allows them to take the
place of forward.

removed forwards. they are incredibly stupid, because they duplicate
claims in almost every way (except forwards had names, claims did not,
which i fixed). now you can have a "Mastodon" claim that points to your
mastodon account etc.

restructing many container types, but mainly claims/stamps:
ClaimContainer no longer exists, and claims/stamps no longer have signed
content. instead, all signing and verification is done at the
transaction level. this means a signed stamp is really just a
normal transaction (which can be included both in the stamper or
stampee's identity). similarly, publishing is no longer a separate
process, but rather a new transaction type (that cannot be saved to the
identity) which allows publishing to also be multisig.

there's a ton of supporting shit for all the above, but that's the main
stuff. a large amount of tests are broken, but the system compiles, and
its time to check things in before i go off on another whim and rewire a
bunch of stuff.