Adding `AdminKeyID` (the admin key version of KeyID)
The idea here is that admin keys are "special" so it makes sense, if
we're referencing them by id, to have a special struct to do so with.
the idea here is that if a function requires referencing a key by id,
and it just so happens we ONLY want an admin key, then we can enforce
this by using `AdminKeyID` instead of just `KeyID`.

added this to our heroic context object, and adding KeyID and AdminKeyID
entries when an admin key is encountered.