Rip out nickname. Allow finding claims by name.
Ripping out nicknames, a stupid and useless feature that could likely
work just as well using the claim system (like we did with forwards). In
fact, most of this shit can just be done with a Url claim and a decent
naming scheme, including nicknames.

In the process, needed to replace the nickname part of the transaction
merge reset test so added a way to find claims by name. last write wins.

will probably end up doing the same with admin keys/subkeys.