Getting rid of StampCollection
instead of the dumbass StampCollection thing, now just embedding
revocations directly in the `Stamp` container as an Option. this makes
it a TON easier when reading the stamps to pick out if it's revoked or
not instead of double-looping two vecs. it just makes the interface
cleaner too. and it's a fairly backwards compatible change as far as the
DAG goes since it deals with StampEntry's directly isntead of Stamps.
win win.