Make Stamp Revocations Great Again
Actually making them great for the first time, because they were pretty
dumb the first time around. They were fine when Stamp wasn't a DAG, but
it has long-since not made sense to have a `StampRevocation` struct and
all the stuff that goes in it.

So instead we're making the Stamp struct have a revocation:
Option<Reason> field, which allows us to quickly iterate stamps and
select active/revoked without cross checking against another list.
Essentially, we're getting rid of the StampCollection and any traces of
it entirely and going with a more natural construction, basically the
exact same one we use in keychain management.