DAG updates!
implemented a new Dag.walk() fn that preserves causal AND timestamp
sorting, not just causal. this walks transactions in order FIRST
causally and then by timestamp (and if timestamps conflict, it uses the
lowest ordered transaction id). this mirrors how the build_identity()
function worked previously because it was the right way of doing things.
however, the walk() implementation is much more efficient because it
makes use of proper data structures (like BTreeSet) instead of doing the
weird sorting shit i was doing before. walk()'s order of visitation now
has its own test as well.

secondly, build_identity() is now full implemented using Dag.walk()!
this is a big deal, because it means Stamp is dogfooding the same DAG
processing utilities that other apps can use to order their
transactions. after finishing the implementation, the main DAG test
(dag::transactions::tests::merge_reset) ran perfectly the first try (not
surprising, a lot of the code was reused).

all of this code is much more readable and commented than it was
previously.