blake2* -> blake3
The API is way nicer and the performance is supposedly better. This is
mostly about the nicer API (ie, not having 200 different variants and
dealing with GenericArray which I hate).

Yes it has less rounds, but Too Much Crypto seems well-reasoned and it's
not like blake3 is obscure so I'm fine with this for now. Also, we can
always throw in Hash::Sha256 if desired.