Making the RNG functional.
Effectively, instead of doing random number generation in-house, the
caller now has to pass in their own RNG for generating keys and such.
This removes a large surface area of non-functionality of Stamp's core
APIs, which also makes testing certain aspects much easier (like the
strict binary serialization format tests).

This is a large change, and marginally obnoxious to users (of which there
is one that I know of: me) but gets us closer to being core-ready and
having std as a feature, making Stamp viable to embedded platforms
eventually. It also means I can start adding binary format stability
tests and really commit to them.

Another benefit is users get more control. They can now seed and supply
their own RNG instead of crossing their fingers and hoping ours is good
enough.

Note that because we can now reliably test serialization, I'm seeing
that rasn (v0.11) is givin different results than the previous versions
when serializing Option<_>'s in a struct (it serializes them as an empty
entry instead of omitting them). I need to consult the DER standard to
determine which one is actually correct, and possibly work with the rasn
team on fixing if it deviates from spec.

For the sanity of users who just don't care at all about RNGs, we export
`crypto::base::rng_chacha20()` to create a CSRNG seeded from the OS'
entropy pool which can be used wherever `&mut rng` is asked for in the
new APIs.

Overall this is a great update. The goal of stamp_core is to be as
functional as possible, and this gets us 99% of the way there. There are
some functions (i think `Timestamp::now()` and `rng_chacha20()` might be
the only ones left...) that remain non-functional, but overall looking
tight as a drum, sir.