Updating DAG walker to use node (transaction) IDs for ancestry instead of u32s.
the idea here is that the u32s will shift around in a graph depending on
which transactions enter the chat at various branch points. this means
if you are depending on the ancestry to be stable, you need stable
identifiers OBVIOUSLY so integers won't cut it.

this took some shifting around and obviously rewriting some tests, but
all's well so far.