Adding Dag.apply()
`Dag.apply()` is a state-building abstraction around `Dag.walk()`. It
walk()s the DAG and applies state as it goes along. it attempts to be as
generic as possible, allowing building of many different types of state.

this completely replaces the `Transactions.build_identity_impl()` fn,
which it inherits most of its code from. *that being said* we fixed some
bugs and bad assumptions in that code and obviously made it much more
generic.

the goal of this is for DAG-walking state builders to be able to
springboard off Dag.apply() instead of implementing the same thing over
and over themselves (as was happening in the stamp net impl).

Also:

- remove orphan error (nobody uses it)
- add txid to TransactionInvalid error, implemented